fix(tools): change .eslintrc rules for ESLint 2 compatibility
ESLint V2.0 removed rules "space-after-keywords" and "space-return-throw-case", it also added the rule "keyword-spacing" that improves upon the previous two separate rules.
This commit removes the aforementioned rules and configures "keyword-spacing" to behave in the exact same way.
The rule is configured to replicate the exact cases that "space-after-keywords" and "space-return-throw-case" covered, however it could be further simplified to cover all cases seen in
http://eslint.org/docs/rules/keyword-spacing

Breaking Changes: Compatibility with ESLint < V2.0