chore(tools): change .eslintrc to be compatible with eslint 2.0
ESLint 2.0 removed the rules "space-after-keywords" and "space-return-throw-case" and replaced them with "keyword-spacing", said rule controls spacing behaviour for spaces before and after keywords, .eslintrc has been changed to reflect that and now spaces are necessary before and after each keyword.

Also, a script was created to select the appropriate .eslintrc file that matches the installed version of eslint, if eslint is not installed globally no file is selected.

Breaking Changes: None